Fix native compilation
The use of nanosleep in the previous job wait change fails to build
for me when building af-main-native, so add the required #include of
time.h.  It's not clear to me right off why it actually built for the
target without it, that behavior difference may warrant some further
investigation.

Bug-AGL: SPEC-3365, SPEC-3427, SPEC-3457

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I73203e564c96d3eaea250ce8867149f6a155ec07